Refactor mysql queries
